Reimplementing the dashboard, fixes #142

+++ src/main/java/de/zalando/ep/zalenium/dashboard/Dashboard.java
@@ -1,6 +1,7 @@
 package de.zalando.ep.zalenium.dashboard;
 
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.collect.Lists;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonObject;
@@ -12,12 +13,17 @@ import org.apache.commons.io.FileUtils;
 
 import java.io.File;
 import java.io.IOException;
+import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
+import org.apache.commons.io.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -34,7 +40,6 @@ public class Dashboard implements DashboardInterface {
     public static final String LOGS_FOLDER_NAME = "logs";
     private static final String TEST_COUNT_FILE = "executedTestsInfo.json";
     private static final String TEST_INFORMATION_FILE = "testInformation.json";
-    private static final String TEST_LIST_FILE = "list.html";
     private static final String DASHBOARD_FILE = "dashboard.html";
     private static final String DASHBOARD_TEMPLATE_FILE = "dashboard_template.html";
     private static final String EXECUTED_TESTS_FIELD = "executedTests";
@@ -42,6 +47,7 @@ public class Dashboard implements DashboardInterface {
     private static final String ZALANDO_ICO = "zalando.ico";
     private static final String CSS_FOLDER = "/css";
     private static final String JS_FOLDER = "/js";
+    private static final String IMG_FOLDER = "/img";
     private static final String ZALENIUM_RETENTION_PERIOD = "ZALENIUM_RETENTION_PERIOD";
     private static final int DEFAULT_RETENTION_PERIOD = 3;
     private static final Logger LOGGER = LoggerFactory.getLogger(Dashboard.class.getName());
@@ -53,8 +59,7 @@ public class Dashboard implements DashboardInterface {
     private static int retentionPeriod;
 
     public Dashboard() {
-        retentionPeriod = env.getIntEnvVariable(ZALENIUM_RETENTION_PERIOD,
-                DEFAULT_RETENTION_PERIOD);
+        retentionPeriod = env.getIntEnvVariable(ZALENIUM_RETENTION_PERIOD, DEFAULT_RETENTION_PERIOD);
     }
 
     public static String getCurrentLocalPath() {
@@ -88,31 +93,9 @@ public class Dashboard implements DashboardInterface {
         
         try {
             synchronizeExecutedTestsValues(testCountFile);
-
-            String testEntry = FileUtils.readFileToString(new File(getCurrentLocalPath(), "list_template.html"), UTF_8);
-            testEntry = testEntry.replace("{fileName}", testInformation.getFileName())
-                    .replace("{testName}", testInformation.getTestName())
-                    .replace("{dateAndTime}", commonProxyUtilities.getShortDateAndTime(testInformation.getTimestamp()))
-                    .replace("{browserAndPlatform}", testInformation.getBrowserAndPlatform())
-                    .replace("{proxyName}", testInformation.getProxyName())
-                    .replace("{seleniumLogFileName}", testInformation.getSeleniumLogFileName())
-                    .replace("{browserDriverLogFileName}", testInformation.getBrowserDriverLogFileName())
-                    .replace("{testStatus}", testInformation.getTestStatus().getTestStatus())
-                    .replace("{testBadge}", testInformation.getTestStatus().getTestBadge())
-                    .replace("{screenDimension}", testInformation.getScreenDimension())
-                    .replace("{timeZone}", testInformation.getTimeZone())
-                    .replace("{testBuild}", testInformation.getBuild())
-                    .replace("{retentionDate}", commonProxyUtilities.getShortDateAndTime(testInformation.getRetentionDate()));
-            
-            File testList = new File(getLocalVideosPath(), TEST_LIST_FILE);
-            // Putting the new entry at the top
-            String testListContents = testEntry;
-            if (testList.exists()) {
-                testListContents = FileUtils.readFileToString(testList, UTF_8);
-                testListContents = testEntry.concat("\n").concat(testListContents);
-            }
-            FileUtils.writeStringToFile(testList, testListContents, UTF_8);
-            CommonProxyUtilities.setFilePermissions(testList.toPath());
+            testInformation.buildSeleniumLogFileName();
+            testInformation.buildBrowserDriverLogFileName();
+            testInformation.setAddedToDashboardTime(new Date().getTime());
 
             executedTests++;
             if (testInformation.isVideoRecorded()) {
@@ -128,12 +111,7 @@ public class Dashboard implements DashboardInterface {
             CommonProxyUtilities.setFilePermissions(testCountFile.toPath());
 
             File dashboardHtml = new File(getLocalVideosPath(), DASHBOARD_FILE);
-            String dashboard = FileUtils.readFileToString(new File(getCurrentLocalPath(), DASHBOARD_TEMPLATE_FILE), UTF_8);
-            dashboard = dashboard.replace("{testList}", testListContents)
-                    .replace("{executedTests}", String.valueOf(executedTests))
-                    .replace("{retentionPeriod}", String.valueOf(retentionPeriod));
-            FileUtils.writeStringToFile(dashboardHtml, dashboard, UTF_8);
-            CommonProxyUtilities.setFilePermissions(dashboardHtml.toPath());
+            setupDashboardFile(dashboardHtml);
 
             File zalandoIco = new File(getLocalVideosPath(), ZALANDO_ICO);
             if (!zalandoIco.exists()) {
@@ -143,6 +121,7 @@ public class Dashboard implements DashboardInterface {
 
             File cssFolder = new File(getLocalVideosPath() + CSS_FOLDER);
             File jsFolder = new File(getLocalVideosPath() + JS_FOLDER);
+            File imgFolder = new File(getLocalVideosPath() + IMG_FOLDER);
 
             if (!cssFolder.exists()) {
                 FileUtils.copyDirectory(new File(getCurrentLocalPath() + CSS_FOLDER), cssFolder);
@@ -151,6 +130,10 @@ public class Dashboard implements DashboardInterface {
             if (!jsFolder.exists()) {
                 FileUtils.copyDirectory(new File(getCurrentLocalPath() + JS_FOLDER), jsFolder);
                 CommonProxyUtilities.setFilePermissions(jsFolder.toPath());
+            }
+            if (!imgFolder.exists()) {
+                FileUtils.copyDirectory(new File(getCurrentLocalPath() + IMG_FOLDER), imgFolder);
+                CommonProxyUtilities.setFilePermissions(imgFolder.toPath());
             }
             saveTestInformation(testInformation);
         } catch (IOException e) {
@@ -168,50 +151,47 @@ public class Dashboard implements DashboardInterface {
         
         if(invalidTestsInformation.size() > 0) {
             LOGGER.info("Cleaning up " + invalidTestsInformation.size() + " test(s) from Dashboard");
-            File testCountFile = new File(getLocalVideosPath(), TEST_COUNT_FILE);
-            File dashboardHtml = new File(getLocalVideosPath(), DASHBOARD_FILE);
-            File testList = new File(getLocalVideosPath(), TEST_LIST_FILE);
-            
+
             for(TestInformation testInformation : invalidTestsInformation) {
                 deleteIfExists(new File(getLocalVideosPath() + "/" + testInformation.getFileName()));
                 deleteIfExists(new File(testInformation.getLogsFolderPath()));
             }
 
-            deleteIfExists(dashboardHtml);
-            deleteIfExists(testList);
-            deleteIfExists(testCountFile);
-
-            String dashboard = FileUtils.readFileToString(new File(getCurrentLocalPath(), DASHBOARD_TEMPLATE_FILE), UTF_8);
-            dashboard = dashboard.replace("{testList}", "").replace("{executedTests}", "0");
-            FileUtils.writeStringToFile(dashboardHtml, dashboard, UTF_8);
-            CommonProxyUtilities.setFilePermissions(dashboardHtml.toPath());
-            
-            for(TestInformation testInformation : validTestsInformation) {
-                updateDashboard(testInformation);
-            }
+            cleanupFiles(false);
             dumpTestInformationToFile(validTestsInformation);
         }
     }
     
     public synchronized void resetDashboard() throws IOException {
         LOGGER.info("Resetting Dashboard");
-        File testList = new File(getLocalVideosPath(), TEST_LIST_FILE);
+        cleanupFiles(true);
+    }
+
+    private void cleanupFiles(boolean reset) throws IOException {
         File testCountFile = new File(getLocalVideosPath(), TEST_COUNT_FILE);
         File dashboardHtml = new File(getLocalVideosPath(), DASHBOARD_FILE);
-        File logsFolder = new File(getLocalVideosPath(), LOGS_FOLDER_NAME);
-        File testInformationFile = new File(getLocalVideosPath(), TEST_INFORMATION_FILE);
-        File videosFolder = new File(getLocalVideosPath());
-        String[] extensions = new String[] { "mp4", "mkv" };
-        for (File file : FileUtils.listFiles(videosFolder, extensions, true)) {
-            deleteIfExists(file);
-        }
-        deleteIfExists(logsFolder);
-        deleteIfExists(testList);
-        deleteIfExists(testCountFile);
+
         deleteIfExists(dashboardHtml);
-        deleteIfExists(testInformationFile);
+        deleteIfExists(testCountFile);
+
+        if (reset) {
+            File logsFolder = new File(getLocalVideosPath(), LOGS_FOLDER_NAME);
+            File testInformationFile = new File(getLocalVideosPath(), TEST_INFORMATION_FILE);
+            File videosFolder = new File(getLocalVideosPath());
+            String[] extensions = new String[] { "mp4", "mkv" };
+            for (File file : FileUtils.listFiles(videosFolder, extensions, true)) {
+                deleteIfExists(file);
+            }
+            deleteIfExists(logsFolder);
+            deleteIfExists(testInformationFile);
+        }
+
+        setupDashboardFile(dashboardHtml);
+    }
+
+    private void setupDashboardFile(File dashboardHtml) throws IOException {
         String dashboard = FileUtils.readFileToString(new File(getCurrentLocalPath(), DASHBOARD_TEMPLATE_FILE), UTF_8);
-        dashboard = dashboard.replace("{testList}", "").replace("{executedTests}", "0");
+        dashboard = dashboard.replace("{retentionPeriod}", String.valueOf(retentionPeriod));
         FileUtils.writeStringToFile(dashboardHtml, dashboard, UTF_8);
         CommonProxyUtilities.setFilePermissions(dashboardHtml.toPath());
     }
@@ -280,6 +260,68 @@ public class Dashboard implements DashboardInterface {
             LOGGER.warn(e.toString(), e);
         }
     }
+
+    public static void saveDashboard() {
+        LOGGER.info("Saving dashboard...");
+        List<TestInformation> executedTestsInformation = loadTestInformationFromFile();
+        executedTestsInformation.sort(Comparator.comparing(TestInformation::getAddedToDashboardTime));
+        String itemTemplate = loadTestItemTemplate();
+        List<String> testItems = new ArrayList<>();
+        for (TestInformation testInformation : executedTestsInformation) {
+            String platformLogo;
+            if (testInformation.getPlatform().toLowerCase().contains("mac")) {
+                platformLogo = "apple";
+            } else if (testInformation.getPlatform().toLowerCase().contains("windows")) {
+                platformLogo = "windows";
+            } else {
+                platformLogo = testInformation.getPlatform().toLowerCase();
+            }
+            String testItem = itemTemplate.replace("{fileName}", testInformation.getFileName())
+                    .replace("{testName}", testInformation.getTestName())
+                    .replace("{seleniumSessionId}", testInformation.getSeleniumSessionId())
+                    .replace("{testStatus}", testInformation.getTestStatus().name())
+                    .replace("{testStatusLowerCase}", testInformation.getTestStatus().name().toLowerCase())
+                    .replace("{browser}", testInformation.getBrowser())
+                    .replace("{browserLowerCase}", testInformation.getBrowser().toLowerCase())
+                    .replace("{browserVersion}", testInformation.getBrowserVersion())
+                    .replace("{platformLogo}", platformLogo)
+                    .replace("{proxyName}", testInformation.getProxyName())
+                    .replace("{proxyNameLowerCase}", testInformation.getProxyName().toLowerCase())
+                    .replace("{timestamp}", testInformation.getTimestamp().toString())
+                    .replace("{addedToDashboardTime}", String.valueOf(testInformation.getAddedToDashboardTime()))
+                    .replace("{screenDimension}", testInformation.getScreenDimension())
+                    .replace("{timeZone}", testInformation.getTimeZone())
+                    .replace("{build}", testInformation.getBuild())
+                    .replace("{seleniumLogFileName}", testInformation.getSeleniumLogFileName())
+                    .replace("{browserDriverLogFileName}", testInformation.getBrowserDriverLogFileName())
+                    .replace("{retentionDate}", testInformation.getRetentionDate().toString());
+            testItems.add(testItem);
+        }
+
+        try {
+            File dashboardHtml = new File(getLocalVideosPath(), DASHBOARD_FILE);
+            String dashboard = FileUtils.readFileToString(new File(getCurrentLocalPath(), DASHBOARD_TEMPLATE_FILE), UTF_8);
+            dashboard = dashboard.replace("list-group\">", String.format("%s%s", "list-group\">",
+                    String.join(System.lineSeparator(), Lists.reverse(testItems))))
+                    .replace("{retentionPeriod}", String.valueOf(retentionPeriod));
+            FileUtils.writeStringToFile(dashboardHtml, dashboard, UTF_8);
+            CommonProxyUtilities.setFilePermissions(dashboardHtml.toPath());
+        } catch (Exception e) {
+            LOGGER.warn(e.toString(), e);
+        }
+    }
+
+    private static String loadTestItemTemplate() {
+        String templateFile = "html_templates/test_item.html";
+        InputStream inputStream = Thread.currentThread().getContextClassLoader().getResourceAsStream(templateFile);
+        try {
+            return IOUtils.toString(Objects.requireNonNull(inputStream), StandardCharsets.UTF_8);
+        } catch (IOException e) {
+            LOGGER.warn(e.toString(), e);
+        }
+        return "";
+    }
+
 
     private void saveTestInformation(TestInformation testInformation) {
         try {
